Fix 3 live-testing bugs: sender verification, search params, Sheets tab name
- config/team.py: EYAL_TELEGRAM_ID falls back to TELEGRAM_EYAL_CHAT_ID
  (inbound filter was rejecting Eyal as unknown sender)
- core/agent.py: fix search_embeddings() param names (match_threshold →
  similarity_threshold, match_count → limit, filter_source_type → source_type)
- services/google_sheets.py: remove hardcoded Sheet1! prefix from all ranges
  (Stakeholder Tracker tab wasn't named "Sheet1", causing parse errors)
- scripts/get_google_token.py: add gmail.modify scope for email watcher
- scripts/setup_supabase.sql: cast ts_rank to FLOAT in fulltext RPC

+++ scripts/setup_supabase.sql
@@ -316,7 +316,7 @@ BEGIN
         e.speaker,
         e.timestamp_range,
         e.metadata,
-        ts_rank(e.chunk_text_tsv, plainto_tsquery('english', search_query)) AS rank
+        ts_rank(e.chunk_text_tsv, plainto_tsquery('english', search_query))::FLOAT AS rank
     FROM embeddings e
     WHERE
         e.chunk_text_tsv @@ plainto_tsquery('english', search_query)
